set ImageAnalysis output format to RGBA_8888
intead of converting yuv to rgba manually, we now let CameraX ImageAnalysis implementation handle that for us -> more efficient

+++ app/src/main/java/com/example/depthcamera/camera/CameraFrameAnalyzer.kt
@@ -64,7 +64,7 @@ class CameraFrameAnalyzer(
 			PerformanceInfo.newCameraFrame()
 
 			val inputBitmap = PerformanceInfo.measureCameraScope("Convert input to bitmap") {
-				image.image!!.toBitmap(image.imageInfo.rotationDegrees)
+				image.image!!.toBitmap(image.imageInfo.rotationDegrees.toFloat())
 			}
 
 			latestCameraFrame.set(inputBitmap)
